git-svn-id: https://svn.lcsr.jhu.edu/cisst/branches/2012-05-04-matlab@3829 0959eb1a-6956-4f60-947b-6092a4e22c97

+++ cisst/cisstOSAbstraction/osaSerialPort.h
@@ -211,7 +211,7 @@ public:
     int Write(const char * data, int nBytes);
     int Write(const unsigned char * data, int nBytes);
     inline int Write(const std::string & data) {
-        return this->Write(data.c_str(), data.size());
+        return this->Write(data.c_str(), static_cast<int>(data.size()));
     }
     
     /*! Receive raw data. */
